refactor(server/seed): Split seed.ts into multiple files
